+++ src/js/renderer/disableSelection.js
@@ -1,0 +1,4 @@
+const path = require('path');
+const insertCssIntoWebFrame = require('./insertCssIntoWebFrame.js');
+
+insertCssIntoWebFrame.fromFile(path.join(__dirname, '../../css/disable-selection.css'));